chore: move and rename files with data (#660)
