+++ app/[lang]/biography/data/BiographyContent.consts.ts
