+++ app/[lang]/biography/data/HeroSection.consts.ts
